dev-db_helper: create functions to insert both associations and recommendations
In functions that insert assocations/recommendations, instead of receiving as parameter the final associative array with all the data, arrays of each field is received. This is because we want to encapsulate all the data logic in this class and, like this, the tables' columns' names are only known in this class. A very little more work at creation, but for an easier maintenance.

Added comment block to "query_data" function, which was missing.