dev-simple_recomm: fix bug creating recomm., previous resources where being recommended (see build nº 130)
It's needed to, first, query the current courses' resources, and, then, associate them to the previous ones, and make the recommendations taking into account these associations. The easiest solution at this moment was to modify the "query_data" function, to allow specifying to take into account the weeks, and also to query only the not viewed contents.