issue #38: leave the first data importation for the first cron execution
Originally, the data import was done in the block itself; now, that operations are in the task that is executed by the cron.
When the task is executed, it cannot know if it is the first time that is being executed for the given course and that if it has to perform the import, so, a database function is needed to check if it is the first task execution for the course.